requirements.txt
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmos==4.6.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
isodate==0.6.1
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
